chore (refactoring): Move BermudaDevices and utils to own files.
